+++ src/Classes/Window.rs
@@ -17,8 +17,8 @@ impl Window {
     // sets pixel color for specified position on window ( x, y )
     pub fn dot(&mut self, point: Point2D, color: &Rgb) {
         let c = color.to_int();
-        let x = point.x - 1;
-        let y = point.y - 1;
+        let x = (point.x - 1.0) as i32;
+        let y = (point.y - 1.0) as i32;
 
         let index = (y * self.width as i32 + x) as usize;
         if x < self.width as i32 && x > 0 && y < self.height as i32 && y > 0 {
@@ -30,10 +30,10 @@ impl Window {
     // draws line from point A to point B
     pub fn line(&mut self, p1: Point2D, p2: Point2D, color: &Rgb) {
 
-        let mut x0 = p1.x;
-        let mut y0 = p1.y;
-        let x1 = p2.x;
-        let y1 = p2.y;
+        let mut x0 = p1.x as i32;
+        let mut y0 = p1.y as i32;
+        let x1 = p2.x as i32;
+        let y1 = p2.y as i32;
 
         let dx = i32::abs(x1 - x0);
         let sx = if x0 < x1 {1} else {-1};
@@ -44,7 +44,7 @@ impl Window {
         loop {
             if x0 == x1 && y0 == y1 {break};
 
-            self.dot(Point2D { x: x0, y: y0 }, color);
+            self.dot(Point2D { x: x0 as f32, y: y0 as f32 }, color);
 
             let e2 = 2 * err;
             if e2 >= dy {
